requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-huggingface>=0.0.7
faiss-cpu>=1.8.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.6.0
tenacity>=8.2.3
tiktoken>=0.5.2
